Pass the parity value itself to get_parity() instead of its position.
For DCF77, there is only one bit array, so we could pass the position unambiguously (or even omit the stop bit position). For NPL, there are two bit arrays and the parity bits are in the other bit array than the value bits.

The check if the parity bit is outside the value range has consequently been removed.